TesteVersion

+++ ola_mundo/lib/utils/temporario_directory.dart
@@ -3,12 +3,12 @@ import 'package:path_provider/path_provider.dart';
 
   Future<void> limparCacheTemp() async{
     try{
-      final dir = await getTemporaryDirectory();
-      final dire = Directory(dir.path);
+      final tempDir = await getTemporaryDirectory();
+      final dire = Directory(tempDir.path);
 
       if(dire.existsSync()){
         await for (var file in dire.list()){
-          if(file is File){
+        if (file is File && file.path.contains('temp_')) {
             await file.delete();
           }
         }
